✨ Add: `impl_for_role!` macro and use it for 3 models [#95]
